style: :lipstick: Designate padding variables for header and main, and update background and logo colors
Designating padding variables for the header and main sections, along with their finalized designs, contributes to a more organized and visually pleasing layout. Additionally, updating the background to include the maindown image and changing the color of the log icon enhances visual consistency and appeal. These updates aim to improve the overall design aesthetic and user experience of the application.